requirements.txt
asgiref==3.11.0
contourpy==1.3.3
cycler==0.12.1
Django==6.0
djangorestframework==3.16.1
fonttools==4.61.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
tzdata==2025.3
